Add Earn > Customer section
