+++ client/my-sites/earn/index.ts
@@ -6,6 +6,7 @@ import earnController from './controller';
 export default function () {
 	page( '/earn', siteSelection, sites, makeLayout, clientRender );
 	page( '/earn/stats', siteSelection, sites, makeLayout, clientRender );
+	page( '/earn/customers', siteSelection, sites, makeLayout, clientRender );
 	page( '/earn/payments', siteSelection, sites, makeLayout, clientRender );
 	// This is legacy, we are leaving it here because it may have been public
 	page(
