+++ client/my-sites/earn/main.tsx
@@ -17,6 +17,7 @@ import { useSelector } from 'calypso/state';
 import { canAccessWordAds, isJetpackSite } from 'calypso/state/sites/selectors';
 import { getSelectedSite } from 'calypso/state/ui/selectors';
 import AdsWrapper from './ads/wrapper';
+import CustomerSection from './customers';
 import Home from './home';
 import MembershipsSection from './memberships';
 import MembershipsProductsSection from './memberships/products';
@@ -52,6 +53,11 @@ const EarningsMain = ( { section, query, path }: EarningsMainProps ) => {
 				title: translate( 'Tools' ),
 				path: '/earn' + pathSuffix,
 				id: 'earn',
+			},
+			{
+				title: translate( 'Customers' ),
+				path: '/earn/customers' + pathSuffix,
+				id: 'customers',
 			},
 			{
 				title: translate( 'Stats' ),
@@ -140,6 +146,9 @@ const EarningsMain = ( { section, query, path }: EarningsMainProps ) => {
 
 			case 'stats':
 				return <StatsSection />;
+
+			case 'customers':
+				return <CustomerSection />;
 
 			case 'refer-a-friend':
 				return <ReferAFriendSection />;
